requirements.txt
llama-index==0.5.13.post1
openai==0.27.4
python-dotenv==0.19.2
